♻️ Refactor HomePage into modular components for better maintainability
- Extract keyboard navigation logic into custom hook (KeyboardNavigation.tsx)
- Create dedicated HeaderActions component for action buttons
- Split project statistics display into ProjectStatsModal component
- Add HelpModal component for keyboard shortcuts display
- Create ProjectCategoryFilter component for category selection
- Reduce HomePage.tsx from ~800 to ~300 lines by extracting reusable components
- Improve code organization and separation of concerns
- Maintain all existing functionality while improving modularity

This refactor makes the homepage more maintainable and testable by breaking down complex logic into focused, reusable components.
